requirements.txt
PyQt5
Pillow
brotli